requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-google-genai
langchain-openai
python-dotenv